finegrained-fp8: fix int32 overflow in fused-grouped expert weight offsets (#1028)
_resolve_tile_inline returned expert_id as int32, so the
expert_id * stride_gu_e / stride_down_e pointer offsets in the fused
grouped kernels overflow at large expert counts x weight strides:
at E=256, H=6144, I=2048 (block-FP8), expert 255's gate_up offset is
255 * 2*I*H = 6.42e9, which wraps to +2.13e9 — silently reading
another expert's weights (or NaN garbage past the region). grouped.py,
batched.py, and fused_batched.py already carry the int64 cast; this
brings fused_grouped.py in line.

Validated on B200: tokens routed to expert 255 at the shape above
returned NaN before the fix and match a torch dequant oracle
(rel 0.043, the normal fused-FP8 band) after.

+++ finegrained-fp8/torch-ext/finegrained_fp8/fused_grouped.py
@@ -107,7 +107,9 @@ def _resolve_tile_inline(
     offs = tl.arange(0, BLOCK_SIZE_M)
     row_mask = offs < freq - within * BLOCK_SIZE_M
     sorted_indices = tl.max_contiguous(m_start + offs, BLOCK_SIZE_M)
-    return expert_id, sorted_indices, row_mask
+    # Cast to int64 so expert_id * stride_gu_e/stride_down_e don't overflow for
+    # large num_experts (matches grouped.py/batched.py; e.g. E=256 x 2*I*H strides).
+    return expert_id.to(tl.int64), sorted_indices, row_mask
 
 
 @triton.jit
